Extract data loading from Chen2021FitnessResults
