+++ src/models/chen2021Fitness/Chen2021FitnessContainer.tsx
@@ -5,6 +5,7 @@ import * as zod from 'zod';
 import { SampleSelectorSchema } from '../../helpers/sample-selector';
 import styled from 'styled-components';
 import { Chen2021FitnessResults } from './Chen2021FitnessResults';
+import { fillRequestWithDefaults } from './loading';
 
 type ContainerProps = zod.infer<typeof SampleSelectorSchema>;
 
@@ -18,19 +19,9 @@ export const Chen2021FitnessContainer = ({
   matchPercentage,
   samplingStrategy,
 }: ContainerProps) => {
-  const [paramData, setParamData] = useState<Chen2021FitnessRequest>({
-    country,
-    mutations,
-    matchPercentage,
-    samplingStrategy,
-    alpha: 0.95,
-    generationTime: 4.8,
-    reproductionNumberWildtype: 1,
-    plotStartDate: new Date('2021-01-01'),
-    plotEndDate: new Date(Date.now() + 60 * 24 * 60 * 60 * 1000),
-    initialWildtypeCases: 1000,
-    initialVariantCases: 100,
-  });
+  const [paramData, setParamData] = useState<Chen2021FitnessRequest>(() =>
+    fillRequestWithDefaults({ country, mutations, matchPercentage, samplingStrategy })
+  );
   const [formGenerationTime, setFormGenerationTime] = useState(paramData.generationTime.toString());
   const [formReproductionNumberWildtype, setFormReproductionNumberWildtype] = useState(
     paramData.reproductionNumberWildtype.toString()
